chore(mvn): move go-offline to profile
Version management of the maven-dependency-plugin should be handled in
the version management of our pom. Therefore, we should not explicitly
call the action from outside but use a profile. In the case of the
go-offline mojo, this profile actually already existed, but was not yet
utilised to execute the plugin's mojo.